darce
Add FFT

+++ SR/SR/SignalBase.cs
@@ -10,7 +10,7 @@ namespace SR
         public int _channels { get; set; }
         public int _length { get; set; }
 
-        public double[] data;
+        public float[] data;
 
         public abstract void Read();
         public abstract void Reset(string fileName = null);
